refactor: JPA 사용으로 인한 쿠폰 도메인의 테스트 코드 수정
